Add a process-global info struct
These are shamefully easy ways of keeping really cross-cutting information
around. I provide both const (immutable) and non-const (mutable) views of the
global struct, and I make the immutable one more enticing by giving it the short
name "g".

In the main() function, we put the argument count and vector in it.